backend/requirements.txt
aiofiles==23.1.0
anyio==3.6.2
fastapi==0.95.1
idna==3.4
Pillow==9.5.0
pydantic==1.10.7
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
